Refactoring: introduce inspecthardware.go
This is a part of the effort to make ironic/ironic.go more manageable.
Currently, it consists of more than 2000 lines and contains most of
the actual provisioning logic.

We already split tests into separate files, I'm thinking of splitting
the code along the same line.
